Refactor project structure: rename Home component import, and remove About, Education, Experience, Footer, ProfileCard, Projects, and Skills components
