+++ src/App.js
@@ -1,4 +1,4 @@
-import Home from "./components/Home";
+import Home from "./components/Home.jsx";
 import "./App.css";
 import { inject } from "@vercel/analytics";
